+++ tests/RestrictionsTest.cpp
@@ -6,7 +6,7 @@
 TEST(RestrictionsTest, basicRestrictions) {
     Restrictions r("aa, bb cc ,");
     std::vector<std::string> restrictions = r.getRestrictions();
-    EXPECT_EQ(restrictions.size(), 3);
+    EXPECT_EQ(restrictions.size(), (size_t)3);
     EXPECT_NE(std::find(restrictions.begin(), restrictions.end(), "aa"), restrictions.end());
     EXPECT_NE(std::find(restrictions.begin(), restrictions.end(), "bb"), restrictions.end());
     EXPECT_NE(std::find(restrictions.begin(), restrictions.end(), "cc"), restrictions.end());
@@ -16,7 +16,7 @@ TEST(RestrictionsTest, emptyRestrictions) {
     Restrictions r1(", , ,  ");
     Restrictions r2("");
     std::vector<std::string> restrictions = r1.getRestrictions();
-    EXPECT_EQ(restrictions.size(), 0);
+    EXPECT_EQ(restrictions.size(), (size_t)0);
     restrictions = r2.getRestrictions();
-    EXPECT_EQ(restrictions.size(), 0);
+    EXPECT_EQ(restrictions.size(), (size_t)0);
 }
